pyproject.toml
[project]
name = "v-flask"
version = "0.1.0"
description = "Flask Core Extension - User, Config, Logging, Auth für Flask-Anwendungen"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Carsten Vogelsang" }
]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-login>=0.6",
    "flask-migrate>=4.0",
    "werkzeug>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
    "pytest-cov>=4.1",
]
markdown = [
    "markdown>=3.5",
]
media = [
    "pillow>=10.0",
    "requests>=2.31",
]
all = [
    "markdown>=3.5",
    "pillow>=10.0",
    "requests>=2.31",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/v_flask", "src/v_flask_plugins"]

[tool.hatch.build.targets.wheel.force-include]
"src/v_flask/templates" = "v_flask/templates"
"src/v_flask/static" = "v_flask/static"
"src/v_flask_plugins/kontakt/templates" = "v_flask_plugins/kontakt/templates"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
]
